+++ src/dfx/src/actors/pocketic.rs
@@ -411,6 +411,8 @@ async fn initialize_pocketic(
             nonmainnet_features: true,
             log_level: Some(replica_config.log_level.to_pocketic_string()),
             bitcoind_addr: bitcoind_addr.clone(),
+            icp_features: None,
+            allow_incomplete_state: None,
         })
         .send()
         .await?