+++ CHANGELOG.md
@@ -6,6 +6,35 @@
 
 * Command `rules` that outputs time-efficient GNU Make rules.
 
+# 0.29.0
+
+### feat: add dfx native support for aarch64-Linux
+
+Add dfx native support for aarch64-Linux.
+
+### feat: support canister snapshot download and upload.
+
+Added `dfx canister snapshot download` and `dfx canister snapshot upload` commands to download and upload the canister snapshot.
+
+## Dependencies
+
+### Motoko
+
+Updated Motoko to [0.15.1](https://github.com/dfinity/motoko/releases/tag/0.15.1)
+
+### Replica
+
+Updated replica to elected commit 615045e039c57ed842c689e49a07ab3de3a8a781.
+This incorporates the following executed proposals:
+
+- [137578](https://dashboard.internetcomputer.org/proposal/137578)
+- [137497](https://dashboard.internetcomputer.org/proposal/137497)
+- [137345](https://dashboard.internetcomputer.org/proposal/137345)
+- [137224](https://dashboard.internetcomputer.org/proposal/137224)
+
+# 0.28.0
+>>>>>>> 92b85e02f3110a0f1d12979634331c66b0ced43c
+
 ### fix: deps deploy works with Canister ID out of the ranges of the pocket-ic subnets
 
 The `dfx deps deploy` command didn't work when the pulled dependency's Canister ID is out of the ranges of the `pocket-ic` subnets.
@@ -40,18 +69,37 @@ You can get the fees by `get_config` API on the [BTC Mainnet Canister](https://d
 If you were using the contents of the `pocket-ic-proxy-port` file to determine the port for
 the `/http_gateway` endpoint, you should instead use `dfx info pocketic-config-port`
 
+### feat: add dfx native support for aarch64-Darwin
+
+Add dfx native support for aarch64-Darwin. Using it may require editing your identities. See the [migration guide](./docs/migration/dfx-0.28.0-migration-guide.md) for more information.
+
 ## Dependencies
 
 ### Replica
 
-Updated replica to commit ac7ff452684f84ea0cfc3fd0a27228220a368b33.
+Updated replica to commit e915efecc8af90993ccfc499721ebe826aadba60.
 This incorporates the following executed proposals:
 
+- [137152](https://dashboard.internetcomputer.org/proposal/137152)
+- [137072](https://dashboard.internetcomputer.org/proposal/137072)
 - [136982](https://dashboard.internetcomputer.org/proposal/136982)
 - [136887](https://dashboard.internetcomputer.org/proposal/136887)
 - [136789](https://dashboard.internetcomputer.org/proposal/136789)
 - [136731](https://dashboard.internetcomputer.org/proposal/136731)
 - [136567](https://dashboard.internetcomputer.org/proposal/136567)
+
+### Motoko
+
+Updated Motoko to [0.14.13](https://github.com/dfinity/motoko/releases/tag/0.14.13)
+
+### Frontend canister
+
+Added missing fields to the asset canister .wasm: `SetAssetContentArguments.last_chunk : opt blob`, and `HttpResponse.upgrade: opt bool`
+
+- Module hash: 2f73b9e18b992f221a5fbab7fc59d840a9cbc461f7cfe875049f51354d23696c
+- https://github.com/dfinity/sdk/pull/4289
+- https://github.com/dfinity/sdk/pull/4285
+- https://github.com/dfinity/sdk/pull/4286
 
 # 0.27.0
 
@@ -81,7 +129,7 @@ Improve `dfx canister logs` with several options
 
 ### Motoko
 
-Updated Motoko to [0.14.11](https://github.com/dfinity/motoko/releases/tag/0.14.11)
+Updated Motoko to [0.14.8](https://github.com/dfinity/motoko/releases/tag/0.14.8)
 
 ### Bitcoin canister